PoT v6/requirements.txt
# PoT Protocol v6 Dependencies

# Core cryptography
liboqs-python>=0.9.0       # SPHINCS+ (FIPS 205), ML-KEM (FIPS 203)
argon2-cffi>=23.1.0        # Argon2id memory-hard PoW
pycryptodome>=3.19.0       # SHA3-256, general crypto

# NTP
ntplib>=0.4.0              # NTP client

# Bitcoin API
httpx>=0.25.0              # Async HTTP client
aiohttp>=3.9.0             # Alternative async HTTP

# Network
websockets>=12.0           # WebSocket support

# Storage
aiosqlite>=0.19.0          # Async SQLite

# Rust bindings (build from source)
maturin>=1.4.0             # Build Rust extensions

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.2.0

# JSON-RPC
# jsonrpcserver>=5.0.0     # Optional, use aiohttp instead

# Optional performance
# uvloop>=0.19.0           # Fast event loop (Unix only)
